Scripts aangepast

+++ Assets/Scripts/Newspaper/Throwing.cs
@@ -14,7 +14,7 @@ public class Throwing : MonoBehaviour
         player = GameObject.Find("Player");
         velX = Mathf.Sin(player.GetComponent<ControlledVelocity>().angleY + 90 * (Mathf.PI / 180)) * Time.deltaTime;
         velZ = Mathf.Cos(player.GetComponent<ControlledVelocity>().angleY + 90 * (Mathf.PI / 180)) * Time.deltaTime;
-        int rigg = GetComponent<Variables>().richting;
+        float rigg = GetComponent<Variables>().richting;
         player = GameObject.Find("Player");
         Debug.Log(player.GetComponent<ControlledVelocity>().velX);
         Debug.Log(player.GetComponent<ControlledVelocity>().velZ);
